Shim thread support
I've forked the discordgo package and implemented enough of the Discord
threading API to be useful. It's not complete, but as far as Popple is
concerned, it's enough to start receiving messages in Threads and acting
on them.

I know this work is in progress in at least one PR against the discordgo
repo, so I will just sit tight on this until they merge their version.
Don't want to step on anyone's toes.